# A little generic poptrie implementation