This branch contains the beginning of an implementation of [Poptrie](http://conferences.sigcomm.org/sigcomm/2015/pdf/papers/p57.pdf). So far it can...

 - build the poptrie lookup structure for prefixes of up to 64 bits
 - look up keys in the poptrie fairly quick using a DynASM based lookup routine
 - be instantiated using a pre-built poptrie structure

Performance on an Intel(R) Xeon(R) CPU E3-1246 v3 @ 3.50GHz (Haswell, Turbo off)

```
PMU analysis (numentries=10000, keysize=32)
build: 0.1290 seconds
lookup: 13217.09 cycles/lookup 28014.35 instructions/lookup
lookup64: 122.94 cycles/lookup 133.22 instructions/lookup
build(direct_pointing): 0.1056 seconds
lookup(direct_pointing): 5519.01 cycles/lookup 11412.01 instructions/lookup
lookup64(direct_pointing): 89.82 cycles/lookup 70.72 instructions/lookup
```

Not implemented yet:

- ~~prefixes/keys of up to 128 bits (i.e., supporting IPv6)~~

Cc @petebristow