@eugeneia 2019-02-05
I have adopted some of the changes suggested by @corsix, which show quite a substantial improvement in performance. I have omitted the loop rotations, since they didn’t seem to make a measurable difference when testing and make the code harder to read imho.

I also didn’t do the BMI2 dependent changes since the oldest lab server doesn’t support them (grindelwald has an ivy bridge CPU), also I would have to add the BMI2 instructions to DynASM first. Might be worth to do a CPU feature detection here?

Before:
```
PMU analysis (numentries=400000, keysize=64)
build: 6.7974 seconds
lookup64: 576.12 cycles/lookup 215.73 instructions/lookup
build(direct_pointing): 7.3970 seconds
lookup64(direct_pointing): 441.36 cycles/lookup 150.59 instructions/lookup
```
After:
```
PMU analysis (numentries=400000, keysize=64)
build: 6.9591 seconds
lookup64: 507.08 cycles/lookup 186.47 instructions/lookup
build(direct_pointing): 7.5430 seconds
lookup64(direct_pointing): 404.69 cycles/lookup 132.36 instructions/lookup
```