@corsix 2019-02-05
FWIW, the latest upstream DynASM should have BMI2 in it (c.f. https://github.com/LuaJIT/LuaJIT/blob/26f1023819efb843e10014232cd88bb1d52ea4f5/dynasm/dasm_x86.lua#L1690, present there since February 2018). Not having Haswell everywhere is a valid reason not to use it though (I would however be interested in knowing what performance improvement - if any - it could give here, if that experiment is something you're willing to try).