@eugeneia 2019-02-06
@petebristow I am having trouble patching in the BMI2 instructions from https://github.com/LuaJIT/LuaJIT/commit/fe651bf6e2b4d02b624be3c289378c08bab2fa9b

I get errors like

```
lib/poptrie_lookup.dasl:98: error: bad char `w' in pattern `F20F38wF7rM' for `shrx':
      | shrx v, key, offset
lib/poptrie_lookup.dasl:109: error: bad char `w' in pattern `0F38wF5rM' for `bzhi':
   | bzhi rax, vec, rcx
lib/poptrie_lookup.dasl:130: error: bad char `w' in pattern `0F38wF5rM' for `bzhi':
   | bzhi rax, vec, rcx
```
 and can’t seem to find where the `w` modifier is actually implemented?