@corsix 2019-02-07
Aha, the `w` got lost as [the PR](https://github.com/LuaJIT/LuaJIT/pull/399/files) became [the upstream commit](https://github.com/LuaJIT/LuaJIT/commit/fe651bf6e2b4d02b624be3c289378c08bab2fa9b). That is unfortunate.