@eugeneia 2019-02-08
@corsix On another note, I am currently updating this code to work with keys up to 128 bits (IPv6) and am wondering if you know any neat tricks for extracting 6 bits from a `uint8_t[16]` at an arbitrary offset in x86 asm? (I suppose what is currently an `uint64_t` argument will have to become a pointer.)