@corsix 2019-02-08
shrd from a pair of uint64_t registers comes to mind. Not a cheap
instruction (e.g. four uops on Broadwell), but should do what you need.

On Fri, 8 Feb 2019 at 08:40, Max Rottenkolber <notifications@github.com>
wrote:

> @corsix <https://github.com/corsix> On another note, I am currently
> updating this code to work with keys up to 128 bits (IPv6) and am wondering
> if you know any neat tricks for extracting 6 bits from a uint8_t[16] at
> an arbitrary offset in x86 asm? (I suppose what is currently an uint64_t
> argument will have to become a pointer.)
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/snabbco/snabb/pull/1397#issuecomment-461729686>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAAGhlWhgvHiBE7hW2JajDkcsGFQxDvJks5vLTf7gaJpZM4ZTsDc>
> .
>