@eugeneia 2019-02-08
Without BMI2 (this branch):
```
PMU analysis (numentries=400000, keysize=64)
build: 6.7322 seconds
lookup: 55733.43 cycles/lookup 87694.39 instructions/lookup
lookup64: 504.83 cycles/lookup 186.46 instructions/lookup
build(direct_pointing): 7.5217 seconds
lookup(direct_pointing): 11152.73 cycles/lookup 25908.96 instructions/lookup
lookup64(direct_pointing): 416.15 cycles/lookup 132.36 instructions/lookup
```
With BMI2 (https://github.com/eugeneia/snabb/commits/poptrie-bmi2):
```
PMU analysis (numentries=400000, keysize=64)
build: 7.3056 seconds
lookup: 39330.65 cycles/lookup 72852.83 instructions/lookup
lookup64: 431.63 cycles/lookup 151.47 instructions/lookup
build(direct_pointing): 7.8106 seconds
lookup(direct_pointing): 12443.39 cycles/lookup 26317.73 instructions/lookup
lookup64(direct_pointing): 376.19 cycles/lookup 112.10 instructions/lookup
```