@eugeneia 2019-02-08
@corsix Thanks for the suggestion! I ended up with something like this for extracting from a 16 byte array, probably not optimal... wdyt?

```
mov rxc, offset
mov rax, [key]
mov rbx, [key+8]
cmp rcx, 63
jng >1
sub rcx, 64
mov rax, rbx
1:
shrd rax, rbx, cl
and rax, 0x3F
```