@corsix 2019-02-08
```
mov ecx, offset   -- Prefer 32-bit operations where possible
mov rax, [key]
mov rbx, [key+8]
test cl, 64       -- Prefer test to cmp (archaic habbit)
cmovnz rax, rbx   -- This is your mov within the branch
                  -- Do not need to sub rcx, 64
                  -- as shrd takes cl mod 64, and subtracting 64 mod 64 is a no-op
shrd rax, rbx, cl
and eax, 0x3F     -- Prefer 32-bit operations where possible
```