@eugeneia 2019-02-11
@corsix Neat! Thanks!

I have added support for 128 (and 32) bit keys (the interface now expects pointers to byte arrays for keys), a well as conditional use of BMI2 instructions (since they have a substantial impact.)

I have adjusted the micro benchmark so that it does not measure a cold cache. It is not exactly representative of a real workload either, but yields more accurate feedback for testing changes. I have also noticed that for 400,000 entries the benchmark will now OOM, but could confirm that this will be fixed by switching to RaptorJIT.

I think the results tell us that its pretty fast with warm caches, and cost increases roughly linearly with key size (prefix length).

```
PMU analysis (numentries=400000, keysize=32)
build: 6.0666 seconds
lookup: 5604.90 cycles/lookup 14653.72 instructions/lookup
lookup32: 81.79 cycles/lookup 118.72 instructions/lookup
lookup64: 81.77 cycles/lookup 118.72 instructions/lookup
lookup128: 99.13 cycles/lookup 138.95 instructions/lookup
build(direct_pointing): 6.0043 seconds
lookup(direct_pointing): 2605.19 cycles/lookup 4805.54 instructions/lookup
lookup32(direct_pointing): 46.62 cycles/lookup 79.66 instructions/lookup
lookup64(direct_pointing): 47.05 cycles/lookup 79.65 instructions/lookup
lookup128(direct_pointing): 51.88 cycles/lookup 84.86 instructions/lookup
```
```
PMU analysis (numentries=400000, keysize=64)
build: 17.5607 seconds
lookup: 48630.60 cycles/lookup 58459.67 instructions/lookup
lookup64: 131.99 cycles/lookup 153.46 instructions/lookup
lookup128: 144.89 cycles/lookup 185.23 instructions/lookup
build(direct_pointing): 17.5885 seconds
lookup(direct_pointing): 4672.99 cycles/lookup 12343.01 instructions/lookup
lookup64(direct_pointing): 92.95 cycles/lookup 114.37 instructions/lookup
lookup128(direct_pointing): 96.92 cycles/lookup 131.13 instructions/lookup
```